requirements.txt
langchain==0.1.8
streamlit==1.31.1
langchain-openai==0.0.8
python-dotenv==1.0.1
openai==1.12.0
beautifulsoup4==4.12.3
langchain_text_splitters==0.0.1
chromadb==0.4.24
langchainhub==0.1.15
langchain-community==0.0.37
langchain-core==0.1.52
sentence-transformers
selenium==4.20.0
openpyxl
bs4
unstructured==0.13.7